Checkbox for labels Mutator

+++ app/Http/Controllers/Admin/ProductController.php
@@ -1,14 +1,11 @@
 <?php
-
 namespace App\Http\Controllers\Admin;
-
 use App\Category;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\ProductRequest;
 use App\Product;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
-
 class ProductController extends Controller
 {
     /**
@@ -21,7 +18,6 @@ class ProductController extends Controller
         $products = Product::get();
         return view('auth.products.index', compact('products'));
     }
-
     /**
      * Show the form for creating a new resource.
      *
@@ -32,37 +28,38 @@ class ProductController extends Controller
         $categories = Category::get();
         return view('auth.products.form', compact('categories'));
     }
-
     /**
      * Store a newly created resource in storage.
      *
-     * @param \Illuminate\Http\Request $request
+     * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
     public function store(ProductRequest $request)
     {
-        $path = $request->file('image')->store('products');
         $params = $request->all();
-        $params['image'] = $path;
+
+        unset($params['image']);
+        if ($request->has('image')) {
+            $params['image'] = $request->file('image')->store('products');
+        }
+
         Product::create($params);
         return redirect()->route('products.index');
     }
-
     /**
      * Display the specified resource.
      *
-     * @param \App\Product $product
+     * @param  \App\Product  $product
      * @return \Illuminate\Http\Response
      */
     public function show(Product $product)
     {
         return view('auth.products.show', compact('product'));
     }
-
     /**
      * Show the form for editing the specified resource.
      *
-     * @param \App\Product $product
+     * @param  \App\Product  $product
      * @return \Illuminate\Http\Response
      */
     public function edit(Product $product)
@@ -70,28 +67,37 @@ class ProductController extends Controller
         $categories = Category::get();
         return view('auth.products.form', compact('product', 'categories'));
     }
-
     /**
      * Update the specified resource in storage.
      *
-     * @param \Illuminate\Http\Request $request
-     * @param \App\Product $product
+     * @param  \Illuminate\Http\Request  $request
+     * @param  \App\Product  $product
      * @return \Illuminate\Http\Response
      */
     public function update(ProductRequest $request, Product $product)
     {
-        Storage::delete($product->image);
-        $path = $request->file('image')->store('products');
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+        if ($request->has('image')) {
+            Storage::delete($product->image);
+            $params['image'] = $request->file('image')->store('products');
+        }
+
+        foreach (['new', 'hit', 'recommend'] as $fieldName) {
+            if (!isset($params[$fieldName])) {
+                $params[$fieldName] = 0;
+            }
+        }
+
+//        dd($params);
+
         $product->update($params);
         return redirect()->route('products.index');
     }
-
     /**
      * Remove the specified resource from storage.
      *
-     * @param \App\Product $product
+     * @param  \App\Product  $product
      * @return \Illuminate\Http\Response
      */
     public function destroy(Product $product)
